requirements.txt
google-analytics-data>=0.18.0
google-analytics-admin>=0.22.0
pandas>=2.0.0
plotly>=5.15.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
openpyxl>=3.1.0
pyarrow>=14.0.0
